docs (README) : 변경사항 수정
