+++ src/main/java/bowling/README.md
@@ -67,7 +67,7 @@
 
 #### Score 패키지
 * FrameScore
-    * 1개 프레임의 최종 합산 점수를 지닌 객체이다.
+    * 1개 프레임의 점수를 지닌 객체이다.
     * 현재 프레임의 점수 결과(Spare Miss 등)에 따라서 점수 계산에 관여하는 추가 투구 회수를 다르다.
     * 추가 투구를 합산하여 점수가 합산되는 Frame의 경우, 기존 객체를 통해 새로운 객체의 점수를 누계하여 생성한다.
     * 점수 계산에 관여하는 투구 회수(leftPitchCounts)가 0일때 계산이 가능하다.
@@ -82,11 +82,9 @@
     * 0 ~ 10 외의 점수로 Score 객체를 가져오려는 경우 예외를 발생시킨다.
     * 10점인지 0점인지를 구분한다.
 
-* ScoreType
+* PitchScoreType
     * 1번의 투구를 통해 얻은 Score가 가질 수 있는 ScoreType(스트라이크, 스페어 등)을 정의한 enum이다.
     * 각 ScoreType별로 고유한 Signature 문자열을 가진다.
-    
-* ScoreFactory
     * ScoreType은 투구(Pitch)의 순서, 즉 이전 투구의 결과에 영향을 받는다.
     * Score를 매개변수로 받아 적합한 ScoreType을 제시해준다.
         * 볼링핀 10개가 새로 주어졌을 때 첫 번째 투구인 경우(마지막 프레임 포함) : initiate
@@ -129,17 +127,21 @@
     * 현재 프레임의 최종 점수가 계산이 가능하면 점수를 반환한다.
     * 만약 스페어나 스트라이크와 같이 다음 프레임의 투구 결과가 점수 합산에 영향을 끼친다면, 다음 프레임에게 계산 작업을 위임한다.
     * 인스턴스 변수로 다음 프레임의 주소를 가진다.
+    * 아직 계산 결과가 도출되지 못하는 상황이면 Optional.empty()를 반환한다.
     
 * FinalFrame
     * 10번 프레임.
     * 2번 투구를 했을 때 스페어나 스트라이크가 없으면 경기 종료.
     * 3번 투구를 마무리한 경우 경기 종료.
+    * 마지막 프레임에서의 스페어나 스트라이크는 별도의 보너스 점수 없이 단순 합만을 제시한다.
+    * 그 외의 로직은 NormalFrame과 동일하다.
 
 * Frames
     * List<Frame> 일급 컬렉션이다.
     * 현재 Play중인 Frame에게 bowl을 요청한다.
     * 다음 프레임으로 넘어갈 수 있는지를 확인한다.
     * 다음 프레임으로 넘어간다.
+    * 현재 확인 가능한 프레임의 숫자 점수 리스트(FramesScores)와, 점수 시그니쳐 리스트(List<String>)을 반환한다.
     
 #### Exception 패키지
 * BowlingBuildingException
